% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{donutMatch}
\alias{donutMatch}
\title{Create a donut chart (match-level)}
\usage{
donutMatch(df, x_lab, y_lab, visCols, themeCol, textCol)
}
\arguments{
\item{df}{A formatted dataframe.}

\item{x_lab}{A string. Labels the x axis.}

\item{y_lab}{A string. Labels the y axis.}

\item{visCols}{A vector of hex color strings. Sets the colors of the chart.}

\item{themeCol}{A hex color string. Sets the background color of the chart.}

\item{textCol}{A hex color string. Sets the text color of the chart legend.}
}
\value{
If all parameters are valid, then the output will be a ggplot object.
}
\description{
donutMatch returns a donut chart ggplot object for individual match stats
}
\details{
This function creates a donut chart visualization of match stats.
}
