% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{donutSznSum}
\alias{donutSznSum}
\title{Create a donut chart (season-level)}
\usage{
donutSznSum(df, plot_lab, visCols, themeCol, textCol)
}
\arguments{
\item{df}{A formatted dataframe.}

\item{plot_lab}{A string. Labels the chart.}

\item{visCols}{A vector of hex color strings. Sets the colors of the chart.}

\item{themeCol}{A hex color string. Sets the background color of the chart.}

\item{textCol}{A hex color string. Sets the text color of the chart legend.}
}
\value{
If all parameters are valid, then the output will be a ggplot object.
}
\description{
donutSznSum returns a donut chart ggplot object for season stats.
}
\details{
This function creates a donut chart visualization of cumulative stat totals for a season.
}
