% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{formatDf}
\alias{formatDf}
\title{Format dataframe}
\usage{
formatDf(
  dt,
  match,
  rowStat,
  statTeam,
  nStats,
  dStats,
  statName,
  xStat,
  valType
)
}
\arguments{
\item{dt}{A dataframe of stats, eg. a Sportscode matrix output.}

\item{match}{(optional) A string. Indicates the game of interest.}

\item{rowStat}{(optional) A string. Indicates the specific row of interest.}

\item{statTeam}{(optional*) A string. Indicates the team for which the user wants to pull the stats.
- *if the user does not pass in a rowStat, they must pass in a statTeam}

\item{nStats}{A vector of strings. Indicate the columns used to calculate the aggregated value, or the numerator of a ratio when dStats are also passed in.}

\item{dStats}{(optional) A vector of strings. Indicate the columns used to calculate the denominator of a ratio.}

\item{statName}{A string. Indicates the formatted stat name for the plot.}

\item{xStat}{(optional) A string. Adds a stat column to the dataframe; for plots that use the stat variable along the x axis.}

\item{valType}{(optional) A string. Indicates how to calculate the stat value and set the stat label for the plot}
}
\value{
If all parameters are valid for the input dataframe, then the output will be a formatted dataframe.
}
\description{
formatDf returns a dataframe formatted to visualize aggregated column stats
}
\details{
This function returns a formatted dataframe for visualizing data when the
column stats are aggregated together (eg. summed, used to calculate a ratio, etc.).
}
