% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{formatDfByCol}
\alias{formatDfByCol}
\title{Format dataframe, by column}
\usage{
formatDfByCol(dt, match, trendTeam, rowStat, colStats, statNames)
}
\arguments{
\item{dt}{A dataframe of stats, eg. a Sportscode matrix output.}

\item{match}{(optional) A string. Indicates the game of interest.}

\item{trendTeam}{(optional) A string. Indicates the team for which the user wants to pull all existing game stats.}

\item{rowStat}{A string. Indicates the row of interest.}

\item{colStats}{A vector of strings. Indicates the columns of interest.}

\item{statNames}{A vector of strings. Indicates the formatted version of the colStat names for the plot.}
}
\value{
If all parameters are valid for the input dataframe, then the output will be a formatted dataframe.
}
\description{
formatDfByCol returns a dataframe formatted to visualize distinct column stats
}
\details{
This function returns a formatted dataframe for visualizing data when the
column stats for analysis each need to be visualized individually.
}
