% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{formatDfByPlayer}
\alias{formatDfByPlayer}
\title{Format dataframe, by player}
\usage{
formatDfByPlayer(dt, match, nStats, dStats, statName, xStat, valType, players)
}
\arguments{
\item{dt}{A dataframe of stats, eg. a Sportscode matrix output.}

\item{match}{(optional) A string. Indicates the game of interest.}

\item{nStats}{A vector of strings. Indicate the columns used to calculate the aggregated value, or the numerator of a ratio when dStats are also passed in.}

\item{dStats}{(optional) A vector of strings. Indicate the columns used to calculate the denominator of a ratio.}

\item{statName}{A string. Indicates the formatted stat name for the plot.}

\item{xStat}{(optional) A string. Adds a stat column to the dataframe; for plots that use the stat variable along the x axis.}

\item{valType}{(optional) A string. Indicates how to calculate the stat value and set the stat label for the plot}

\item{players}{A vector of strings. Specifies the names of the players for whom to pull stats.}
}
\value{
If all parameters are valid for the input dataframe, then the output will be a formatted dataframe.
}
\description{
formatDfByPlayer returns a dataframe formatted to visualize stats across players
}
\details{
This function returns a formatted dataframe for visualizing data when
comparing stats across players
}
