% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{formatDfComplexRatio}
\alias{formatDfComplexRatio}
\title{Format dataframe, complex ratios}
\usage{
formatDfComplexRatio(
  dt,
  match,
  trendTeam,
  nrowStat,
  ncolStats,
  drowStat,
  dcolStats,
  statNames,
  ratioLab
)
}
\arguments{
\item{dt}{A dataframe of stats, eg. a Sportscode matrix output.}

\item{match}{(optional) A string. Indicates the game of interest.}

\item{trendTeam}{(optional) A string. Indicates the team for which the user wants to pull all existing game stats.}

\item{nrowStat}{A string. Indicates the row of interest for the numerator stats.}

\item{ncolStats}{A vector of strings. Indicates the columns used to calculate the numerator of the ratio.}

\item{drowStat}{A string. Indicates the row of interest for the denominator stats.}

\item{dcolStats}{A vector of strings. Indicates the columns used to calculate the denominator of the ratio.}

\item{statNames}{A vector of strings. Indicates the formatted numerator/denominator stat names for the plot.}

\item{ratioLab}{A string. Indicates the name of calculated ratio for the plot.}
}
\value{
If all parameters are valid for the input dataframe, then the output will be a formatted dataframe.
}
\description{
formatDfComplexRatio returns a dataframe formatted to visualize multi-row stat ratios
}
\details{
This function returns a formatted dataframe for visualizing data when the
ratio being calculated is calculated from different row stats.
}
