% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{groupStatMatch}
\alias{groupStatMatch}
\title{Create a bar chart grouped by stat (match-level)}
\usage{
groupStatMatch(df, x_lab, y_lab, visCols)
}
\arguments{
\item{df}{A formatted dataframe.}

\item{x_lab}{A string. Labels the x axis.}

\item{y_lab}{A string. Labels the y axis.}

\item{visCols}{A vector of hex color strings. Sets the colors of the chart.}
}
\value{
If all parameters are valid, then the output will be a ggplot object.
}
\description{
groupStatMatch returns a bar chart, grouped by stat, ggplot object for individual match stats
}
\details{
This function creates a grouped bar chart visualization comparing team/opp match stats,
where color is based on the stat.
}
