% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{stackGroupTrend}
\alias{stackGroupTrend}
\title{Create a multi-panel stacked bar chart (season-level)}
\usage{
stackGroupTrend(df, x_lab, y_lab, visCols)
}
\arguments{
\item{df}{A formatted dataframe.}

\item{x_lab}{A string. Labels the x axis.}

\item{y_lab}{A string. Labels the y axis.}

\item{visCols}{A vector of hex color strings. Sets the colors of the chart.}
}
\value{
If all parameters are valid, then the output will be a ggplot object.
}
\description{
stackGroupTrend returns a multi-panel stacked bar chart ggplot object for season stats
}
\details{
This function creates a multi-panel stacked bar chart visualization of a stat trend over time, grouped by game.
}
