% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{stackMatch}
\alias{stackMatch}
\title{Create a stacked bar chart (match-level)}
\usage{
stackMatch(df, x_lab, y_lab, visCols, y_lim)
}
\arguments{
\item{df}{A formatted dataframe.}

\item{x_lab}{A string. Labels the x axis.}

\item{y_lab}{A string. Labels the y axis.}

\item{visCols}{A vector of hex color strings. Sets the colors of the chart.}

\item{y_lim}{(optional) A number to set the max y value for the chart. Standardizes the visualization scales across the team/opp.}
}
\value{
If all parameters are valid, then the output will be a ggplot object.
}
\description{
stackMatch returns a stacked bar chart ggplot object for individual match stats
}
\details{
This function creates a stacked bar chart visualization comparing team/opp match stats.
column stats are aggregated together (eg. summed, used to calculate a ratio, etc.).
}
