% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{totalFacetTrend}
\alias{totalFacetTrend}
\title{Create a multi-panel line chart (season-level)}
\usage{
totalFacetTrend(df, x_lab, y_lab, visCols)
}
\arguments{
\item{df}{A formatted datframe. The Stat variable in the df is the stat by which the function plots a new panel (e.g. per player).}

\item{x_lab}{A string. Labels the x axis.}

\item{y_lab}{A string. Labels the y axis.}

\item{visCols}{A vector of hex color strings. Sets the colors of the chart.}
}
\value{
If all parameters are valid, then the output will be a ggplot object.
}
\description{
totalFacetTrend returns a multi-panel line chart ggplot object for season stats
}
\details{
This function creates a multi-panel line chart visualization of a stat trend over time.
}
